move base class for system dependencies into base.py
In accordance with review comments; it's small enough this seems fitting.
