+++ mesonbuild/modules/python.py
@@ -31,7 +31,7 @@ from ..build import known_shmod_kwargs
 from .. import mlog
 from ..environment import detect_cpu_family
 from ..dependencies import DependencyMethods, PkgConfigDependency, NotFoundDependency
-from ..dependencies.system import SystemDependency
+from ..dependencies.base import SystemDependency
 from ..programs import ExternalProgram, NonExistingExternalProgram
 
 mod_kwargs = {'subdir'}
